feat(multi-lang-csv-prompt-picker): stored helper functions in own sub-services to reduce code cluttering and create general helper services (toast.js)
